split out gate component
